FAIT-Calculators/HomeHealthScore/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.0
openai==0.27.6
supabase==1.0.3
stripe==5.4.0
elevenlabs==0.2.18
pytest==7.3.1
httpx==0.24.0
pytest-asyncio==0.21.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
